change price to double

+++ Models/Sale.cs
@@ -15,7 +15,7 @@ namespace Labs.Models
         public int id_car { get; set; }
         public DateTime date2 { get; set; }
         public DateTime date3 { get; set; }
-        public int summ { get; set; }
+        public double summ { get; set; }
         public string status { get; set; }
         public int? id_payment { get; set; }
